@@ -3,7 +3,7 @@ module it.auties.protobuf.parser {
     requires java.compiler;
 
     exports it.auties.protobuf.parser;
-    exports it.auties.protobuf.parser.statement;
     exports it.auties.protobuf.parser.exception;
     exports it.auties.protobuf.parser.type;
+    exports it.auties.protobuf.parser.tree;
 }
