@@ -2,26 +2,21 @@ package it.auties.protobuf.serialization;
 
 import com.sun.source.tree.*;
 import com.sun.source.util.Trees;
-import it.auties.protobuf.annotation.ProtobufBuilder;
-import it.auties.protobuf.annotation.ProtobufConverter;
-import it.auties.protobuf.annotation.ProtobufEnumIndex;
-import it.auties.protobuf.annotation.ProtobufProperty;
+import it.auties.protobuf.annotation.*;
 import it.auties.protobuf.extension.OptionalExtension;
 import it.auties.protobuf.model.ProtobufEnum;
 import it.auties.protobuf.model.ProtobufMessage;
 import it.auties.protobuf.model.ProtobufObject;
 import it.auties.protobuf.model.ProtobufType;
-import it.auties.protobuf.serialization.converter.ProtobufConverterElement;
 import it.auties.protobuf.serialization.converter.ProtobufDeserializerElement;
 import it.auties.protobuf.serialization.converter.ProtobufSerializerElement;
 import it.auties.protobuf.serialization.instrumentation.ProtobufDeserializationVisitor;
 import it.auties.protobuf.serialization.instrumentation.ProtobufSerializationVisitor;
-import it.auties.protobuf.serialization.message.ProtobufBuilderElement;
-import it.auties.protobuf.serialization.message.ProtobufEnumMetadata;
-import it.auties.protobuf.serialization.message.ProtobufMessageElement;
-import it.auties.protobuf.serialization.property.ProtobufPropertyStub;
+import it.auties.protobuf.serialization.object.ProtobufBuilderElement;
+import it.auties.protobuf.serialization.object.ProtobufEnumMetadata;
+import it.auties.protobuf.serialization.object.ProtobufMessageElement;
 import it.auties.protobuf.serialization.property.ProtobufPropertyType;
-import it.auties.protobuf.serialization.util.PropertyUtils;
+import it.auties.protobuf.serialization.property.ProtobufPropertyStub;
 import it.auties.protobuf.stream.ProtobufInputStream;
 import it.auties.protobuf.stream.ProtobufOutputStream;
 
@@ -34,6 +29,8 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.lang.annotation.Annotation;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
@@ -47,11 +44,14 @@ import java.util.stream.IntStream;
 @SupportedSourceVersion(SourceVersion.RELEASE_17)
 public class ProtobufJavacPlugin extends AbstractProcessor {
     private Trees trees;
-    private DeclaredType objectType;
     private DeclaredType protoObjectType;
     private DeclaredType protoMessageType;
+    private DeclaredType protoMixinType;
+    private DeclaredType objectType;
     private DeclaredType protoEnumType;
     private DeclaredType collectionType;
+    private DeclaredType mapType;
+    private DeclaredType concurrentMapType;
     private DeclaredType listType;
     private TypeElement arrayListTypeElement;
     private DeclaredType setType;
@@ -59,6 +59,8 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     private DeclaredType queueType;
     private DeclaredType dequeType;
     private TypeElement linkedListTypeElement;
+    private TypeElement hashMapElement;
+    private TypeElement concurrentHashMapElement;
     private PrimitiveType intType;
     private Map<String, ProtobufSerializerElement> optionalSerializers;
     private Map<String, ProtobufDeserializerElement> optionalDeserializers;
@@ -76,16 +78,19 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     }
 
     private void initTypes() {
-        this.objectType = (DeclaredType) getType(Object.class);
         this.protoObjectType = (DeclaredType) getType(ProtobufObject.class);
         this.protoMessageType = (DeclaredType) getType(ProtobufMessage.class);
         this.protoEnumType = (DeclaredType) getType(ProtobufEnum.class);
+        this.objectType = (DeclaredType) getType(Object.class);
         this.intType = (PrimitiveType) getType(int.class);
         this.collectionType = (DeclaredType) erase(getType(Collection.class));
+        this.mapType = (DeclaredType) erase(getType(Map.class));
+        this.concurrentMapType = (DeclaredType) erase(getType(ConcurrentMap.class));
         this.listType = (DeclaredType) erase(getType(List.class));
         var arrayListType = (DeclaredType) erase(getType(ArrayList.class));
         this.arrayListTypeElement = (TypeElement) arrayListType.asElement();
         this.setType = (DeclaredType) erase(getType(Set.class));
+        this.protoMixinType = (DeclaredType) erase(getType(ProtobufMixin.class));
         var hashSetType = (DeclaredType) erase(getType(HashSet.class));
         this.arrayListTypeElement = (TypeElement) arrayListType.asElement();
         this.hashSetTypeElement = (TypeElement) hashSetType.asElement();
@@ -93,6 +98,10 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
         this.dequeType = (DeclaredType) erase(getType(Deque.class));
         var linkedListType = (DeclaredType) erase(getType(LinkedList.class));
         this.linkedListTypeElement = (TypeElement) linkedListType.asElement();
+        var hashMapType = (DeclaredType) erase(getType(HashMap.class));
+        this.hashMapElement = (TypeElement) hashMapType.asElement();
+        var concurrentHashMapType = (DeclaredType) erase(getType(ConcurrentHashMap.class));
+        this.concurrentHashMapElement = (TypeElement) concurrentHashMapType.asElement();
     }
 
     private void initAtomicConverters() {
@@ -112,11 +121,11 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                 AtomicReference.class.getName(),
                 new ProtobufSerializerElement(atomicReferenceSerializer, false, false),
                 AtomicInteger.class.getName(),
-                new ProtobufSerializerElement(atomicIntegerSerializer, false, false),
+                new ProtobufSerializerElement(atomicIntegerSerializer, true, false),
                 AtomicLong.class.getName(),
-                new ProtobufSerializerElement(atomicLongSerializer, false, false),
+                new ProtobufSerializerElement(atomicLongSerializer, true, false),
                 AtomicBoolean.class.getName(),
-                new ProtobufSerializerElement(atomicBooleanSerializer, false, false)
+                new ProtobufSerializerElement(atomicBooleanSerializer, true, false)
         );
         this.atomicDeserializers = Map.of(
                 AtomicReference.class.getName(),
@@ -211,6 +220,12 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                 protoMessageType,
                 ProtobufProperty.class,
                 "Illegal enclosing class: a field annotated with @ProtobufProperty should be enclosed by a class or record that implements ProtobufMessage"
+        );
+        checkEnclosing(
+                roundEnv,
+                protoMessageType,
+                ProtobufGetter.class,
+                "Illegal enclosing class: a method annotated with @ProtobufGetter should be enclosed by a class or record that implements ProtobufMessage"
         );
         checkEnclosing(
                 roundEnv,
@@ -353,14 +368,13 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
 
             writer.println("public class %s {".formatted(simpleGeneratedClassName));
             var invocationArgs = new ArrayList<String>();
-            List<ProtobufPropertyStub> properties = messageElement.properties();
             if(builderElement != null) {
                 for(var parameter : builderElement.parameters()) {
                     writer.println("    private %s %s;".formatted(parameter.asType(), parameter.getSimpleName()));
                     invocationArgs.add(parameter.getSimpleName().toString());
                 }
             }else {
-                for (var property : properties) {
+                for (var property : messageElement.properties()) {
                     writer.println("    private %s %s;".formatted(property.type().fieldType(), property.name()));
                     invocationArgs.add(property.name());
                 }
@@ -368,8 +382,8 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
             writer.println();
             writer.println("    public %s() {".formatted(simpleGeneratedClassName));
             if(builderElement == null) {
-                for (var property : properties) {
-                    writer.println("        %s = %s;".formatted(property.name(), PropertyUtils.getPropertyDefaultValue(property)));
+                for (var property : messageElement.properties()) {
+                    writer.println("        %s = %s;".formatted(property.name(), property.defaultValue()));
                 }
             }
 
@@ -380,7 +394,7 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                     writeBuilderSetter(writer, parameter.getSimpleName().toString(), parameter.asType(), simpleGeneratedClassName);
                 }
             }else {
-                for(var property : properties) {
+                for(var property : messageElement.properties()) {
                     writeBuilderSetter(writer, property.name(), property.type().fieldType(), simpleGeneratedClassName);
                 }
             }
@@ -452,11 +466,12 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
 
     private Optional<ProtobufMessageElement> processMessage(TypeElement message) {
         var messageElement = new ProtobufMessageElement(message, null);
-        processMessage(messageElement);
+        processMessage(messageElement, messageElement.element());
         if(messageElement.properties().isEmpty()) {
             printWarning("No properties found", message);
             return Optional.of(messageElement);
         }
+
 
         if (!hasPropertiesConstructor(messageElement)) {
             printError("Missing protobuf constructor: a protobuf message must provide a constructor that takes only its properties, following their declaration order, as parameters", message);
@@ -467,21 +482,41 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     }
 
 
-    private void processMessage(ProtobufMessageElement messageElement) {
-        for (var entry : messageElement.element().getEnclosedElements()) {
-            if (entry instanceof VariableElement variableElement) {
-                processProperty(messageElement, variableElement);
-            }else if(entry instanceof ExecutableElement executableElement) {
-                var builder = executableElement.getAnnotation(ProtobufBuilder.class);
-                if(builder != null) {
-                    messageElement.addBuilder(builder.className(), executableElement.getParameters(), executableElement);
-                }
+    private void processMessage(ProtobufMessageElement messageElement, TypeElement typeElement) {
+        getSuperClass(typeElement)
+                .ifPresent(superClass -> processMessage(messageElement, superClass));
+        for (var entry : typeElement.getEnclosedElements()) {
+            switch (entry) {
+                case VariableElement variableElement -> processProperty(messageElement, variableElement);
+                case ExecutableElement executableElement -> processMethod(messageElement, executableElement);
+                case null, default -> {}
             }
         }
     }
 
+    private Optional<TypeElement> getSuperClass(TypeElement typeElement) {
+        var superClass = typeElement.getSuperclass();
+        if(superClass == null || superClass.getKind() == TypeKind.NONE) {
+            return Optional.empty();
+        }
+
+        var superClassName = superClass.toString();
+        var superClassTypeParametersStartIndex = superClassName.indexOf("<");
+        var rawName = superClassTypeParametersStartIndex <= 0 ? superClassName : superClassName.substring(0, superClassTypeParametersStartIndex);
+        var superClassElement = processingEnv.getElementUtils().getTypeElement(rawName);
+        return Optional.ofNullable(superClassElement);
+    }
+
+    private void processMethod(ProtobufMessageElement messageElement, ExecutableElement executableElement) {
+        var builder = executableElement.getAnnotation(ProtobufBuilder.class);
+        if(builder != null) {
+            messageElement.addBuilder(builder.className(), executableElement.getParameters(), executableElement);
+        }
+    }
+
     private boolean hasPropertiesConstructor(ProtobufMessageElement message) {
-        return message.element().getEnclosedElements()
+        return message.element()
+                .getEnclosedElements()
                 .stream()
                 .filter(entry -> entry instanceof ExecutableElement)
                 .map(entry -> (ExecutableElement) entry)
@@ -492,7 +527,6 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                         return false;
                     }
 
-                    // No var args support as it's needed
                     var propertiesIterator = message.properties().iterator();
                     var constructorParametersIterator = constructorParameters.iterator();
                     while (propertiesIterator.hasNext() && constructorParametersIterator.hasNext()) {
@@ -521,12 +555,7 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
            return;
         }
 
-        if(propertyAnnotation.repeated() && !checkRepeatedField(variableElement)) {
-            return;
-        }
-
-
-        var accessor = getAccessor(variableElement).orElse(null);
+        var accessor = getAccessor(variableElement, propertyAnnotation).orElse(null);
         if(accessor == null) {
             printError("Missing getter/accessor: a getter or accessor must be declared", variableElement);
             return;
@@ -537,7 +566,7 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
             return;
         }
 
-        var type = getImplementationType(variableElement, accessor, propertyAnnotation);
+        var type = getImplementationType(variableElement, variableElement.asType(), accessor, propertyAnnotation);
         if(type.isEmpty()) {
             return;
         }
@@ -550,116 +579,247 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
         printError("Duplicated message property: %s and %s with index %s".formatted(variableElement.getSimpleName(), error.get().name(), propertyAnnotation.index()), variableElement);
     }
 
-    private Optional<ProtobufPropertyType> getImplementationType(VariableElement field, ExecutableElement accessor, ProtobufProperty property) {
-        var implementationAstType = getExplicitImplementationType(field, property);
-        if (property.repeated()) {
-            var collectionTypeParameter = getCollectionTypeParameter(field.asType())
-                    .orElse(null);
-            if (collectionTypeParameter == null) {
-                printError("Type inference error: cannot determine collection's type. Specify the implementation explicitly in @ProtobufProperty", field);
+    private Optional<ProtobufPropertyType> getImplementationType(Element element, TypeMirror elementType, Element accessor, ProtobufProperty property) {
+        var mixin = getMirroredMixin(property).orElse(null);
+        var override = getMirroredOverride(property);
+        if (isSubType(elementType, collectionType)) {
+            var collectionTypeParameter = getTypeParameter(elementType, collectionType, 0);
+            if (collectionTypeParameter.isEmpty()) {
+                printError("Type inference error: cannot determine collection's implementationType. Specify the implementation explicitly in @ProtobufProperty", element);
                 return Optional.empty();
             }
 
-            var concreteCollectionType = getConcreteCollectionType(field)
-                    .orElse(null);
-            if(concreteCollectionType == null) {
-                printError("Type inference error: no known implementation found for abstract type. Use a concrete type for the property's field to fix this error", field);
+            var concreteCollectionType = getConcreteCollectionType(elementType);
+            if(concreteCollectionType.isEmpty()) {
+                printError("Type inference error: no known implementation found for abstract implementationType. Use a concrete type for the property's field to fix this error", element);
                 return Optional.empty();
             }
 
-            var actualCollectionTypeParameter = implementationAstType.orElse(collectionTypeParameter);
-            var result = new ProtobufPropertyType(property.type(), field.asType(), actualCollectionTypeParameter, concreteCollectionType, false, isEnum(actualCollectionTypeParameter));
-            var converters = getPropertyConverters(field, implementationAstType.orElse(null), actualCollectionTypeParameter, property.type());
-            converters.forEach(result::addNullableConverter);
-            return Optional.of(result);
+            var value = new ProtobufPropertyType.NormalType(property.type(), collectionTypeParameter.get(), override.orElse(collectionTypeParameter.get()), isEnum(collectionTypeParameter.get()));
+            var type = new ProtobufPropertyType.CollectionType(elementType, concreteCollectionType.get(), value);
+            attributeConverters(element, property.type(), collectionTypeParameter.get(), value, mixin);
+            return Optional.of(type);
         }
 
-        var rawAccessorType = erase(accessor.getReturnType());
-        var rawFieldType = erase(field.asType());
+        if(isSubType(elementType, mapType)) {
+            var implementation = getConcreteMapImplementation(element, elementType, property.keyType(), property.valueType());
+            if (implementation.isEmpty()) {
+                printError("Type inference error: no known implementation found for abstract implementationType. Use a concrete implementationType for the property's field to fix this error", element);
+                return Optional.empty();
+            }
+
+            attributeConverters(element, implementation.get().keyType().protobufType(), implementation.get().keyType().fieldType(), implementation.get().keyType(), mixin);
+            attributeConverters(element, implementation.get().valueType().protobufType(), implementation.get().valueType().fieldType(), implementation.get().valueType(), mixin);
+            return Optional.of(implementation.get());
+        }
+
+        var rawAccessorType = switch (accessor) {
+            case ExecutableElement executableElement -> erase(executableElement.getReturnType());
+            case VariableElement variableElement -> erase(variableElement.asType());
+            default -> throw new IllegalStateException("Unexpected value: " + accessor);
+        };
+        var rawFieldType = erase(elementType);
         var optionalDeserializer = optionalDeserializers.get(rawFieldType.toString());
         if (optionalDeserializer != null) {
-            var declaredFieldType = (DeclaredType) field.asType();
+            var declaredFieldType = (DeclaredType) elementType;
             var wrappedType = getOptionalValueType(declaredFieldType);
             if (wrappedType.isEmpty()) {
                 return Optional.empty();
             }
 
             var rawWrappedType = erase(wrappedType.get());
-            var result = new ProtobufPropertyType(property.type(), field.asType(), implementationAstType.orElse(rawWrappedType), null, isOptional(field.asType()), isEnum(rawWrappedType));
-            result.addNullableConverter(optionalSerializers.get(rawAccessorType.toString()));
-            var converters = getPropertyConverters(field, implementationAstType.orElse(null), rawWrappedType, property.type());
-            converters.forEach(result::addNullableConverter);
-            result.addNullableConverter(optionalDeserializer);
-            return Optional.of(result);
+            var value = new ProtobufPropertyType.NormalType(property.type(), wrappedType.get(), override.orElse(wrappedType.get()), isEnum(wrappedType.get()));
+            var implementation = new ProtobufPropertyType.OptionalType(elementType, value);
+            value.addNullableConverter(optionalSerializers.get(rawAccessorType.toString()));
+            attributeConverters(element, property.type(), rawWrappedType, value, mixin);
+            value.addNullableConverter(optionalDeserializer);
+            return Optional.of(implementation);
         }
 
         var atomicDeserializer = atomicDeserializers.get(rawFieldType.toString());
         if (atomicDeserializer != null) {
-            var declaredFieldType = (DeclaredType) field.asType();
+            var declaredFieldType = (DeclaredType) elementType;
             var wrappedType = getAtomicValueType(declaredFieldType);
             if (wrappedType.isEmpty()) {
                 return Optional.empty();
             }
 
             var rawWrappedType = erase(wrappedType.get());
-            var result = new ProtobufPropertyType(property.type(), field.asType(), implementationAstType.orElse(rawWrappedType), null, isOptional(field.asType()), isEnum(rawWrappedType));
-            result.addNullableConverter(atomicSerializers.get(rawAccessorType.toString()));
-            var converters = getPropertyConverters(field, implementationAstType.orElse(null), rawWrappedType, property.type());
-            converters.forEach(result::addNullableConverter);
-            result.addNullableConverter(atomicDeserializer);
-            return Optional.of(result);
+            var value = new ProtobufPropertyType.NormalType(property.type(), wrappedType.get(), override.orElse(wrappedType.get()), isEnum(wrappedType.get()));
+            var implementation = new ProtobufPropertyType.AtomicType(elementType, value);
+            value.addNullableConverter(atomicSerializers.get(rawAccessorType.toString()));
+            attributeConverters(element, property.type(), rawWrappedType, value, mixin);
+            value.addNullableConverter(atomicDeserializer);
+            return Optional.of(implementation);
         }
 
-        var result = new ProtobufPropertyType(property.type(), field.asType(), implementationAstType.orElse(field.asType()), null, isOptional(field.asType()), isEnum(rawFieldType));
-        result.addNullableConverter(optionalSerializers.get(rawAccessorType.toString()));
-        result.addNullableConverter(atomicSerializers.get(rawAccessorType.toString()));
-        var converters = getPropertyConverters(field, implementationAstType.orElse(null), field.asType(), property.type());
-        converters.forEach(result::addNullableConverter);
-        return Optional.of(result);
+        var implementation = new ProtobufPropertyType.NormalType(property.type(), elementType, override.orElse(elementType), isEnum(rawFieldType));
+        implementation.addNullableConverter(optionalSerializers.get(rawAccessorType.toString()));
+        implementation.addNullableConverter(atomicSerializers.get(rawAccessorType.toString()));
+        attributeConverters(element, property.type(), elementType, implementation, mixin);
+        return Optional.of(implementation);
     }
 
-    private static Optional<ExecutableElement> getAccessor(VariableElement fieldElement) {
-        var fieldName = fieldElement.getSimpleName().toString();
-        return fieldElement.getEnclosingElement()
+    private Optional<Element> getMirroredMixin(ProtobufProperty property) {
+        try {
+            var implementation = property.mixin();
+            var type = processingEnv.getElementUtils().getTypeElement(implementation.getName());
+            if(type == null || isSameType(type.asType(), protoMixinType)) {
+                return Optional.empty();
+            }
+
+            return Optional.of(type);
+        }catch (MirroredTypeException exception) {
+            if (!(exception.getTypeMirror() instanceof DeclaredType declaredType)) {
+                return Optional.empty();
+            }
+
+            if(isSameType(declaredType, protoMixinType)) {
+                return Optional.empty();
+            }
+
+            return Optional.of(declaredType.asElement());
+        }
+    }
+
+    private Optional<TypeMirror> getMirroredOverride(ProtobufProperty property) {
+        try {
+            var implementation = property.overrideType();
+            var type = processingEnv.getElementUtils().getTypeElement(implementation.getName());
+            if(type == null || isSameType(type.asType(), objectType)) {
+                return Optional.empty();
+            }
+
+            return Optional.of(type.asType());
+        }catch (MirroredTypeException exception) {
+            if (!(exception.getTypeMirror() instanceof DeclaredType declaredType)) {
+                return Optional.empty();
+            }
+
+            if(isSameType(declaredType, objectType)) {
+                return Optional.empty();
+            }
+
+            return Optional.of(declaredType);
+        }
+    }
+
+    private Optional<? extends Element> getAccessor(VariableElement fieldElement, ProtobufProperty propertyAnnotation) {
+        if(!fieldElement.getModifiers().contains(Modifier.PRIVATE)) {
+            return Optional.of(fieldElement);
+        }
+
+        var methods = fieldElement.getEnclosingElement()
                 .getEnclosedElements()
                 .stream()
                 .filter(entry -> entry instanceof ExecutableElement)
                 .map(entry -> (ExecutableElement) entry)
-                .filter(entry -> {
-                    var methodName = entry.getSimpleName().toString();
-                    return entry.getParameters().isEmpty()
-                            && (methodName.equalsIgnoreCase("get" + fieldName) || methodName.equalsIgnoreCase(fieldName));
-                })
+                .toList();
+        return methods.stream()
+                .filter(entry -> isProtobufGetter(entry, propertyAnnotation))
+                .findFirst()
+                .or(() -> inferAccessor(fieldElement, methods));
+    }
+
+    private Optional<ExecutableElement> inferAccessor(VariableElement fieldElement, List<ExecutableElement> methods) {
+        var fieldName = fieldElement.getSimpleName().toString();
+        return methods.stream()
+                .filter(entry -> isProtobufGetter(entry, fieldName))
                 .findFirst();
     }
 
-    private Optional<TypeMirror> getConcreteCollectionType(VariableElement fieldElement) {
-        if(!(fieldElement.asType() instanceof DeclaredType declaredFieldType)) {
+    private boolean isProtobufGetter(ExecutableElement entry, String fieldName) {
+        var methodName = entry.getSimpleName().toString();
+        return entry.getParameters().isEmpty() && (methodName.equalsIgnoreCase("get" + fieldName) || methodName.equalsIgnoreCase(fieldName));
+    }
+
+    private boolean isProtobufGetter(ExecutableElement entry, ProtobufProperty propertyAnnotation) {
+        var annotation = entry.getAnnotation(ProtobufGetter.class);
+        return annotation != null && annotation.index() == propertyAnnotation.index();
+    }
+
+    private Optional<TypeMirror> getConcreteCollectionType(TypeMirror elementType) {
+        if(!(elementType instanceof DeclaredType declaredFieldType)) {
             return Optional.empty();
         }
 
-
         var fieldTypeElement = (TypeElement) declaredFieldType.asElement();
         if(!fieldTypeElement.getModifiers().contains(Modifier.ABSTRACT)) {
-            return Optional.of(fieldElement.asType());
+            return Optional.of(elementType);
         }
 
-        var rawFieldType = erase(fieldElement.asType());
         var typeArguments = declaredFieldType.getTypeArguments() != null ? declaredFieldType.getTypeArguments().toArray(TypeMirror[]::new) : new TypeMirror[0];
-        if(isSameType(rawFieldType, collectionType) || isSameType(rawFieldType, listType)) {
+        if(isSameType(elementType, collectionType) || isSameType(elementType, listType)) {
             return Optional.ofNullable(processingEnv.getTypeUtils().getDeclaredType(arrayListTypeElement, typeArguments));
-        }else if(isSameType(rawFieldType, setType)) {
+        }else if(isSameType(elementType, setType)) {
             return Optional.ofNullable(processingEnv.getTypeUtils().getDeclaredType(hashSetTypeElement, typeArguments));
-        }else if(isSameType(rawFieldType, queueType) || isSameType(rawFieldType, dequeType)) {
+        }else if(isSameType(elementType, queueType) || isSameType(elementType, dequeType)) {
             return Optional.ofNullable(processingEnv.getTypeUtils().getDeclaredType(linkedListTypeElement, typeArguments));
         }else {
             return Optional.empty();
         }
     }
 
+    private Optional<ProtobufPropertyType.MapType> getConcreteMapImplementation(Element element, TypeMirror elementType, ProtobufType keyType, ProtobufType valueType) {
+        if(!(elementType instanceof DeclaredType declaredFieldType)) {
+            return Optional.empty();
+        }
+
+        var error = false;
+        var keyTypeParameter = getTypeParameter(elementType, mapType, 0);
+        if (keyTypeParameter.isEmpty()) {
+            printError("Type inference error: cannot determine map's key implementationType. Specify the implementation explicitly in @ProtobufProperty", element);
+            error = true;
+        }
+
+        if(keyType == ProtobufType.MAP) {
+            printError("Missing implementationType error: specify the implementationType of the map's key in @ProtobufProperty", element);
+            error = true;
+        }
+
+        if(valueType == ProtobufType.MAP) {
+            printError("Missing implementationType error: specify the implementationType of the map's value in @ProtobufProperty", element);
+            error = true;
+        }
+
+        if(keyType == ProtobufType.OBJECT) {
+            printError("Type error: unsupported map key implementationType", element);
+            error = true;
+        }
+
+        if(error) {
+            return Optional.empty();
+        }
+
+        var keyEntry = new ProtobufPropertyType.NormalType(keyType, keyTypeParameter.get(), keyTypeParameter.get(), false);
+        var valueTypeParameter = getTypeParameter(elementType, mapType, 1);
+        if (valueTypeParameter.isEmpty()) {
+            printError("Type inference error: cannot determine map's value implementationType. Specify the implementation explicitly in @ProtobufProperty", element);
+            return Optional.empty();
+        }
+
+        var valueEntry = new ProtobufPropertyType.NormalType(valueType, valueTypeParameter.get(), valueTypeParameter.get(), isEnum(valueTypeParameter.get()));
+        var fieldTypeElement = (TypeElement) declaredFieldType.asElement();
+        if(!fieldTypeElement.getModifiers().contains(Modifier.ABSTRACT)) {
+            return Optional.of(new ProtobufPropertyType.MapType(elementType, elementType, keyEntry, valueEntry));
+        }
+
+        var rawFieldType = erase(elementType);
+        var typeArguments = declaredFieldType.getTypeArguments() != null ? declaredFieldType.getTypeArguments().toArray(TypeMirror[]::new) : new TypeMirror[0];
+        if(isSameType(rawFieldType, mapType)) {
+            var hashMapType = processingEnv.getTypeUtils().getDeclaredType(hashMapElement, typeArguments);
+            return Optional.of(new ProtobufPropertyType.MapType(elementType, hashMapType, keyEntry, valueEntry));
+        }else if(isSameType(rawFieldType, concurrentMapType)) {
+            var concurrentHashMapType = processingEnv.getTypeUtils().getDeclaredType(concurrentHashMapElement, typeArguments);
+            return Optional.of(new ProtobufPropertyType.MapType(elementType, concurrentHashMapType, keyEntry, valueEntry));
+        } else {
+            return Optional.empty();
+        }
+    }
+
     private Optional<TypeMirror> getAtomicValueType(DeclaredType declaredFieldType) {
         return switch (declaredFieldType.asElement().getSimpleName().toString()) {
-            case "AtomicReference" -> declaredFieldType.getTypeArguments().isEmpty() ? Optional.empty() : Optional.of(declaredFieldType.getTypeArguments().get(0));
+            case "AtomicReference" -> declaredFieldType.getTypeArguments().isEmpty() ? Optional.empty() : Optional.of(declaredFieldType.getTypeArguments().getFirst());
             case "AtomicInteger" -> Optional.of(getType(Integer.class));
             case "AtomicLong" -> Optional.of(getType(Long.class));
             case "AtomicBoolean" -> Optional.of(getType(Boolean.class));
@@ -670,7 +830,7 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     private Optional<TypeMirror> getOptionalValueType(DeclaredType declaredFieldType) {
         return switch (declaredFieldType.asElement().getSimpleName().toString()) {
             case "Optional" ->
-                    declaredFieldType.getTypeArguments().isEmpty() ? Optional.empty() : Optional.of(declaredFieldType.getTypeArguments().get(0));
+                    declaredFieldType.getTypeArguments().isEmpty() ? Optional.empty() : Optional.of(declaredFieldType.getTypeArguments().getFirst());
             case "OptionalLong" -> Optional.of(getType(Long.class));
             case "OptionalInt" -> Optional.of(getType(Integer.class));
             case "OptionalDouble" -> Optional.of(getType(Double.class));
@@ -678,53 +838,44 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
         };
     }
 
-    private List<ProtobufConverterElement> getPropertyConverters(VariableElement invoker, TypeMirror implementation, TypeMirror fieldType, ProtobufType protobufType) {
-        if (implementation == null) {
-            var wrapperType = getType(protobufType.wrappedType());
-            return isSubType(fieldType, wrapperType) ? List.of() : getConvertersForType(invoker, fieldType, wrapperType);
-
+    private void attributeConverters(Element invoker, ProtobufType from, TypeMirror to, ProtobufPropertyType implementation, Element mixin) {
+        var fromType = getType(from.wrappedType());
+        if(!(to instanceof DeclaredType declaredType)) {
+            return;
         }
 
-        return isSubType(fieldType, implementation) ? List.of() : getConvertersForType(invoker, implementation, fieldType);
-    }
+        if(!isSubType(to, fromType) || (from == ProtobufType.OBJECT && !isSubType(to, protoMessageType) && !isSubType(to, protoEnumType))) {
+            ExecutableElement serializer = null;
+            ExecutableElement deserializer = null;
+            var convertersWrapper = Objects.requireNonNullElse(mixin, declaredType.asElement());
+            for (var entry : convertersWrapper.getEnclosedElements()) {
+                if (serializer != null && deserializer != null) {
+                    break;
+                }
 
-    private List<ProtobufConverterElement> getConvertersForType(VariableElement invoker, TypeMirror to, TypeMirror from) {
-        if (!(to instanceof DeclaredType declaredType)) {
-            return List.of();
-        }
+                if (!(entry instanceof ExecutableElement element)) {
+                    continue;
+                }
 
-        ExecutableElement serializer = null;
-        ExecutableElement deserializer = null;
-        for (var entry : declaredType.asElement().getEnclosedElements()) {
-            if(serializer != null && deserializer != null) {
-                break;
+                if (isDeserializer(element, fromType)) {
+                    deserializer = element;
+                } else if (isSerializer(element, to, fromType)) {
+                    serializer = element;
+                }
             }
 
-            if (!(entry instanceof ExecutableElement element)) {
-                continue;
+            if (serializer != null) {
+                implementation.addNullableConverter(new ProtobufSerializerElement(serializer, serializer.getReturnType().getKind().isPrimitive(), isOptional(serializer)));
+            } else {
+                printError("Missing converter: cannot find a serializer for %s".formatted(to), invoker);
             }
 
-            if (isDeserializer(element, from)) {
-                deserializer = element;
-            }else if(isSerializer(element, from)){
-                serializer = element;
+            if (deserializer != null) {
+                implementation.addNullableConverter(new ProtobufDeserializerElement(deserializer));
+            } else {
+                printError("Missing converter: cannot find a deserializer for %s".formatted(fromType), invoker);
             }
         }
-
-        var results = new ArrayList<ProtobufConverterElement>();
-        if(serializer != null) {
-            results.add(new ProtobufSerializerElement(serializer, serializer.getReturnType().getKind().isPrimitive(), isOptional(serializer)));
-        }else {
-            printError("Missing converter: cannot find a element for %s".formatted(from), invoker);
-        }
-
-        if(deserializer != null) {
-            results.add(new ProtobufDeserializerElement(deserializer));
-        }else {
-            printError("Missing converter: cannot find a deserializer for %s".formatted(from), invoker);
-        }
-
-        return Collections.unmodifiableList(results);
     }
 
     private boolean isOptional(ExecutableElement serializer) {
@@ -736,47 +887,19 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
         return optionalSerializers.get(erased.toString()) != null;
     }
 
-    private boolean isDeserializer(ExecutableElement entry, TypeMirror fieldAstType) {
+    private boolean isDeserializer(ExecutableElement entry, TypeMirror from) {
         return entry.getAnnotation(ProtobufConverter.class) != null
+                && entry.getModifiers().contains(Modifier.STATIC)
                 && entry.getParameters().size() == 1
-                && isSubType(fieldAstType, entry.getParameters().get(0).asType());
+                && isSubType(from, entry.getParameters().getFirst().asType());
     }
 
-    private boolean isSerializer(ExecutableElement entry, TypeMirror fieldAstType) {
+    private boolean isSerializer(ExecutableElement entry, TypeMirror to, TypeMirror from) {
+        var isStatic = entry.getModifiers().contains(Modifier.STATIC);
         return entry.getAnnotation(ProtobufConverter.class) != null
-                && entry.getParameters().isEmpty()
-                && isSubType(fieldAstType, entry.getReturnType());
-    }
-
-    private Optional<TypeMirror> getExplicitImplementationType(VariableElement element, ProtobufProperty property) {
-        var rawImplementation = getMirroredImplementation(property);
-        if(rawImplementation.isEmpty()) {
-            return Optional.empty();
-        }
-
-        if (isSameType(rawImplementation.get(), objectType)) {
-            return Optional.empty();
-        }
-
-        if(rawImplementation.get().asElement().getModifiers().contains(Modifier.ABSTRACT)) {
-            printError("Type error: a concrete implementation is required", element);
-            return Optional.empty();
-        }
-
-        return Optional.of(rawImplementation.get());
-    }
-
-    private Optional<DeclaredType> getMirroredImplementation(ProtobufProperty property) {
-        try {
-            var implementation = property.implementation();
-            throw new IllegalArgumentException(implementation.toString());
-        }catch (MirroredTypeException exception) {
-            if(exception.getTypeMirror() instanceof DeclaredType declaredType) {
-                return Optional.of(declaredType);
-            }
-
-            return Optional.empty();
-        }
+                && entry.getParameters().size() == (isStatic ? 1 : 0)
+                && (!isStatic || isSubType(to, entry.getParameters().getFirst().asType()))
+                && isSubType(from, entry.getReturnType());
     }
 
     private boolean isEnum(TypeMirror mirror) {
@@ -784,9 +907,14 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                 && declaredType.asElement().getKind() == ElementKind.ENUM;
     }
 
-    private Optional<TypeMirror> getCollectionTypeParameter(TypeMirror mirror) {
+    private Optional<TypeMirror> getTypeParameter(TypeMirror mirror, TypeMirror targetType, int index) {
         if(!(mirror instanceof DeclaredType declaredType)) {
             return Optional.empty();
+        }
+
+        if (isSameType(mirror, targetType)) {
+            var collectionTypeArgument = declaredType.getTypeArguments().get(index);
+            return getConcreteTypeParameter(collectionTypeArgument, declaredType, index);
         }
 
         var typeElement = (TypeElement) declaredType.asElement();
@@ -794,50 +922,49 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
                 .stream()
                 .filter(implemented -> implemented instanceof DeclaredType)
                 .map(implemented -> (DeclaredType) implemented)
-                .map(implemented -> getCollectionTypeByImplement(declaredType, implemented))
+                .map(implemented -> getTypeParameterByImplement(declaredType, implemented, targetType, index))
                 .flatMap(Optional::stream)
                 .findFirst()
-                .or(() -> getCollectionTypeBySuperClass(declaredType, typeElement));
+                .or(() -> getTypeParameterBySuperClass(declaredType, typeElement, targetType, index));
     }
 
-    private Optional<TypeMirror> getCollectionTypeByImplement(DeclaredType declaredType, DeclaredType implemented) {
-        if (isSameType(implemented, collectionType)) {
-            var collectionTypeArgument = implemented.getTypeArguments().get(0);
-            return getConcreteCollectionArgumentType(collectionTypeArgument, declaredType);
+    private Optional<TypeMirror> getTypeParameterByImplement(DeclaredType declaredType, DeclaredType implemented, TypeMirror targetType, int index) {
+        if (isSameType(implemented, targetType)) {
+            var collectionTypeArgument = implemented.getTypeArguments().get(index);
+            return getConcreteTypeParameter(collectionTypeArgument, declaredType, index);
         }
 
-        return getCollectionTypeParameter(implemented)
-                .flatMap(result -> getConcreteCollectionArgumentType(result, declaredType));
+        return getTypeParameter(implemented, targetType, index)
+                .flatMap(result -> getConcreteTypeParameter(result, declaredType, index));
     }
 
-    private Optional<TypeMirror> getCollectionTypeBySuperClass(DeclaredType declaredType, TypeElement typeElement) {
+    private Optional<TypeMirror> getTypeParameterBySuperClass(DeclaredType declaredType, TypeElement typeElement, TypeMirror targetType, int index) {
         if (!(typeElement.getSuperclass() instanceof DeclaredType superDeclaredType)) {
             return Optional.empty();
         }
 
-        return getCollectionTypeParameter(superDeclaredType)
-                .flatMap(result -> getConcreteCollectionArgumentType(result, superDeclaredType))
-                .flatMap(result -> getConcreteCollectionArgumentType(result, declaredType));
+        return getTypeParameter(superDeclaredType, targetType, index)
+                .flatMap(result -> getConcreteTypeParameter(result, superDeclaredType, index))
+                .flatMap(result -> getConcreteTypeParameter(result, declaredType, index));
     }
 
-    private Optional<TypeMirror> getConcreteCollectionArgumentType(TypeMirror argumentMirror, DeclaredType previousType) {
-        if(argumentMirror instanceof DeclaredType declaredTypeArgument) {
-            return Optional.of(declaredTypeArgument);
-        }else if(argumentMirror instanceof TypeVariable typeVariableArgument){
-            return getConcreteTypeFromTypeVariable(typeVariableArgument, previousType);
-        }else {
-            return Optional.empty();
-        }
+    private Optional<TypeMirror> getConcreteTypeParameter(TypeMirror argumentMirror, DeclaredType previousType, int index) {
+        return switch (argumentMirror) {
+            case DeclaredType declaredTypeArgument -> Optional.of(declaredTypeArgument);
+            case ArrayType arrayType -> Optional.of(arrayType);
+            case TypeVariable typeVariableArgument -> getConcreteTypeFromTypeVariable(typeVariableArgument, previousType, index);
+            case null, default -> Optional.empty();
+        };
     }
 
-    private Optional<TypeMirror> getConcreteTypeFromTypeVariable(TypeVariable typeVariableArgument, DeclaredType previousType) {
+    private Optional<TypeMirror> getConcreteTypeFromTypeVariable(TypeVariable typeVariableArgument, DeclaredType previousType, int index) {
         var currentTypeVarName = typeVariableArgument.asElement().getSimpleName();
         var previousTypeArguments = previousType.getTypeArguments();
         var previousElement = (TypeElement) previousType.asElement();
         var previousTypeParameters = previousElement.getTypeParameters();
-        for(var i = 0; i < previousTypeParameters.size() && i < previousTypeArguments.size(); i++) {
-            if(previousTypeParameters.get(i).getSimpleName().equals(currentTypeVarName)){
-                return Optional.of(previousTypeArguments.get(i));
+        for(;index < previousTypeParameters.size() && index < previousTypeArguments.size(); index++) {
+            if(previousTypeParameters.get(index).getSimpleName().equals(currentTypeVarName)){
+                return Optional.of(previousTypeArguments.get(index));
             }
         }
         return Optional.empty();
@@ -848,20 +975,8 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     }
 
     private TypeMirror erase(TypeMirror typeMirror) {
-        if(!(typeMirror instanceof DeclaredType declaredType)) {
-            return typeMirror;
-        }
-
-        return processingEnv.getTypeUtils().erasure(declaredType);
-    }
-
-    private boolean checkRepeatedField(VariableElement variableElement) {
-        if (isSubType(variableElement.asType(), collectionType)) {
-            return true;
-        }
-
-        printError("Type mismatch: the type of a repeated field must extend java.lang.Collection", variableElement);
-        return false;
+        var result = processingEnv.getTypeUtils().erasure(typeMirror);
+        return result == null ? typeMirror : result;
     }
 
     private boolean isSubType(TypeMirror child, TypeMirror parent) {
@@ -888,17 +1003,12 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
     }
 
     private boolean isValidPackedProperty(VariableElement variableElement, ProtobufProperty propertyAnnotation) {
-        if(!propertyAnnotation.repeated()) {
-            printError("Only repeated properties can be packed", variableElement);
-            return false;
+        if(!propertyAnnotation.packed() || isSubType(variableElement.asType(), collectionType)) {
+            return true;
         }
 
-        if(!propertyAnnotation.type().isPackable()) {
-            printError("Only scalar properties can be packed", variableElement);
-            return false;
-        }
-
-        return true;
+        printError("Only scalar properties can be packed", variableElement);
+        return false;
     }
 
     private Optional<ProtobufMessageElement> processEnum(TypeElement enumElement) {
@@ -966,7 +1076,7 @@ public class ProtobufJavacPlugin extends AbstractProcessor {
 
     private boolean isImplicitEnumConstructor(ExecutableElement constructor) {
         return constructor.getParameters().size() == 1
-                && isSameType(constructor.getParameters().get(0).asType(), intType);
+                && isSameType(constructor.getParameters().getFirst().asType(), intType);
     }
 
     private boolean hasProtobufIndexAnnotation(ExecutableElement constructor, int index) {
