@@ -1,9 +1,9 @@
 package it.auties.protobuf.serialization.instrumentation;
 
 import it.auties.protobuf.model.ProtobufType;
-import it.auties.protobuf.serialization.message.ProtobufMessageElement;
+import it.auties.protobuf.serialization.object.ProtobufMessageElement;
+import it.auties.protobuf.serialization.property.ProtobufPropertyType;
 import it.auties.protobuf.serialization.property.ProtobufPropertyStub;
-import it.auties.protobuf.serialization.util.PropertyUtils;
 
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.TypeElement;
@@ -13,6 +13,8 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 public class ProtobufDeserializationVisitor extends ProtobufInstrumentationVisitor {
+    private static final String DEFAULT_STREAM_NAME = "inputStream";
+
     public ProtobufDeserializationVisitor(ProtobufMessageElement element, PrintWriter writer) {
         super(element, writer);
     }
@@ -72,26 +74,25 @@ public class ProtobufDeserializationVisitor extends ProtobufInstrumentationVisit
         writer.println("            return null;");
         writer.println("        }");
         // ProtobufInputStream stream = new ProtobufInputStream(var1);
-        writer.println("        var inputStream = new ProtobufInputStream(input);");
+        writer.println("        var %s = new ProtobufInputStream(input);".formatted(DEFAULT_STREAM_NAME));
 
-        // [<type> var<index> = <defaultValue>, ...]
+        // [<implementationType> var<index> = <defaultValue>, ...]
         for(var property : message.properties()) {
-            var defaultValue = PropertyUtils.getPropertyDefaultValue(property);
-            writer.println("        %s %s = %s;".formatted(property.type().fieldType(), property.name(), defaultValue));
+            writer.println("        %s %s = %s;".formatted(property.type().fieldType(), property.name(), property.defaultValue()));
         }
 
         // while(input.readTag())
-        writer.println("        while(inputStream.readTag()) {");
+        writer.println("        while(%s.readTag()) {".formatted(DEFAULT_STREAM_NAME));
 
         // switch(input.index())
-        writer.println("            switch(inputStream.index()) {");
+        writer.println("            switch(%s.index()) {".formatted(DEFAULT_STREAM_NAME));
         var argumentsList = new ArrayList<String>();
         for(var property : message.properties()) {
-            var readMethod = getDeserializerStreamMethod(property);
-            var readValue = getReadValue(property, readMethod);
-            var readFunction = getConvertedValue(property, readValue);
-            var readAssignment = getReadAssignment(property, readFunction);
-            writer.println("                case %s -> %s;".formatted(property.index(), readAssignment));
+            switch (property.type()) {
+                case ProtobufPropertyType.MapType mapType -> writeMapSerializer(property, mapType);
+                case ProtobufPropertyType.CollectionType collectionType -> writeDeserializer(property.name(), property.index(), collectionType.value(), true, property.packed());
+                default -> writeDeserializer(property.name(), property.index(), property.type(), false, property.packed());
+            }
             argumentsList.add(property.name());
         }
         writer.println("                default -> inputStream.skipBytes();");
@@ -108,8 +109,40 @@ public class ProtobufDeserializationVisitor extends ProtobufInstrumentationVisit
         writer.println("        return new %s(%s);".formatted(message.element(), String.join(", ", argumentsList)));
     }
 
+    private void writeMapSerializer(ProtobufPropertyStub property, ProtobufPropertyType.MapType mapType) {
+        writer.println("                case %s -> {".formatted(property.index()));
+        var streamName = "%sInputStream".formatted(property.name());
+        var keyName = "%sKey".formatted(property.name());
+        var valueName = "%sValue".formatted(property.name());
+        writer.println("                        var %s = new ProtobufInputStream(%s.readBytes());".formatted(streamName, DEFAULT_STREAM_NAME));
+        writer.println("                        %s %s = null;".formatted(mapType.keyType().fieldType(), keyName));
+        writer.println("                        %s %s = null;".formatted(mapType.valueType().fieldType(), valueName));
+        var keyReadMethod = getDeserializerStreamMethod(mapType.keyType(), false);
+        var keyReadValue = getReadValue(streamName, mapType.keyType(), keyReadMethod);
+        var keyReadFunction = getConvertedValue(mapType.keyType(), keyReadValue);
+        var valueReadMethod = getDeserializerStreamMethod(mapType.valueType(), false);
+        var valueReadValue = getReadValue(streamName, mapType.valueType(), valueReadMethod);
+        var valueReadFunction = getConvertedValue(mapType.valueType(), valueReadValue);
+        writer.println("                        while(%s.readTag()) {".formatted(streamName));
+        writer.println("                            switch(%s.index()) {".formatted(streamName));
+        writer.println("                                case 1 -> %s = %s;".formatted(keyName, keyReadFunction));
+        writer.println("                                case 2 -> %s = %s;".formatted(valueName, valueReadFunction));
+        writer.println("                            }");
+        writer.println("                        }");
+        writer.println("                        %s.put(%s, %s);".formatted(property.name(), keyName, valueName));
+        writer.println("                }");
+    }
+
+    private void writeDeserializer(String name, int index, ProtobufPropertyType type, boolean repeated, boolean packed) {
+        var readMethod = getDeserializerStreamMethod(type, packed);
+        var readValue = getReadValue(DEFAULT_STREAM_NAME, type, readMethod);
+        var readFunction = getConvertedValue(type, readValue);
+        var readAssignment = getReadAssignment(name, repeated, packed, readFunction);
+        writer.println("                case %s -> %s;".formatted(index, readAssignment));
+    }
+
     private void checkRequiredProperty(ProtobufPropertyStub property) {
-        if (!property.repeated()) {
+        if (!(property.type() instanceof ProtobufPropertyType.CollectionType)) {
             writer.println("        Objects.requireNonNull(%s, \"Missing required property: %s\");".formatted(property.name(), property.name()));
             return;
         }
@@ -118,32 +151,32 @@ public class ProtobufDeserializationVisitor extends ProtobufInstrumentationVisit
         writer.println("            throw new NullPointerException(\"Missing required property: %s\");".formatted(property.name()));
     }
 
-    private String getReadAssignment(ProtobufPropertyStub property, String readFunction) {
-        if (!property.repeated()) {
-            return "%s = %s".formatted(property.name(), readFunction);
+    private String getReadAssignment(String name, boolean repeated, boolean packed, String readFunction) {
+        if (!repeated) {
+            return "%s = %s".formatted(name, readFunction);
         }
 
-        var repeatedMethod = property.packed() ? "addAll" : "add";
-        return "%s.%s(%s)".formatted(property.name(), repeatedMethod, readFunction);
+        var repeatedMethod = packed ? "addAll" : "add";
+        return "%s.%s(%s)".formatted(name, repeatedMethod, readFunction);
     }
 
-    private String getReadValue(ProtobufPropertyStub property, String readMethod) {
-        var reader = "inputStream.%s()".formatted(readMethod);
-        if (property.type().protobufType() != ProtobufType.OBJECT) {
+    private String getReadValue(String streamName, ProtobufPropertyType type, String readMethod) {
+        var reader = "%s.%s()".formatted(streamName, readMethod);
+        if (type.protobufType() != ProtobufType.OBJECT) {
             return reader;
         }
 
-        var specName = getSpecName(property.type().implementationType());
-        if(property.type().isEnum()) {
+        var specName = getSpecName(type.implementationType());
+        if(type.isEnum()) {
             return "%s.decode(%s).orElse(null)".formatted(specName, reader);
         }
 
         return "%s.decode(%s)".formatted(specName, reader);
     }
 
-    private String getConvertedValue(ProtobufPropertyStub property, String readValue) {
+    private String getConvertedValue(ProtobufPropertyType implementation, String readValue) {
         var result = readValue;
-        for(var converter : property.type().deserializers()) {
+        for(var converter : implementation.deserializers()) {
             if (converter.element().getKind() == ElementKind.CONSTRUCTOR) {
                 var converterWrapperClass = (TypeElement) converter.element().getEnclosingElement();
                 result = "new %s(%s)".formatted(converterWrapperClass.getQualifiedName(), result);
@@ -157,17 +190,22 @@ public class ProtobufDeserializationVisitor extends ProtobufInstrumentationVisit
     }
 
     // Returns the method to use to deserialize a property from ProtobufInputStream
-    private String getDeserializerStreamMethod(ProtobufPropertyStub property) {
-        return property.type().isEnum() ? property.packed() ? "readInt32Packed" : "readInt32" : switch (property.type().protobufType()) {
+    private String getDeserializerStreamMethod(ProtobufPropertyType type, boolean packed) {
+        if (type.isEnum()) {
+            return packed ? "readInt32Packed" : "readInt32";
+        }
+        
+        return switch (type.protobufType()) {
             case STRING -> "readString";
             case OBJECT, BYTES -> "readBytes";
-            case BOOL -> property.packed() ? "readBoolPacked" : "readBool";
-            case INT32, SINT32, UINT32 -> property.packed() ? "readInt32Packed" : "readInt32";
-            case FLOAT -> property.packed() ? "readFloatPacked" : "readFloat";
-            case DOUBLE -> property.packed() ? "readDoublePacked" : "readDouble";
-            case FIXED32, SFIXED32 -> property.packed() ? "readFixed32Packed" : "readFixed32";
-            case INT64, SINT64, UINT64 -> property.packed() ? "readInt64Packed" : "readInt64";
-            case FIXED64, SFIXED64 -> property.packed() ? "readFixed64Packed" : "readFixed64";
+            case BOOL -> packed ? "readBoolPacked" : "readBool";
+            case INT32, SINT32, UINT32 -> packed ? "readInt32Packed" : "readInt32";
+            case FLOAT -> packed ? "readFloatPacked" : "readFloat";
+            case DOUBLE -> packed ? "readDoublePacked" : "readDouble";
+            case FIXED32, SFIXED32 -> packed ? "readFixed32Packed" : "readFixed32";
+            case INT64, SINT64, UINT64 -> packed ? "readInt64Packed" : "readInt64";
+            case FIXED64, SFIXED64 -> packed ? "readFixed64Packed" : "readFixed64";
+            default -> throw new IllegalStateException("Unexpected value: " + type.protobufType());
         };
     }
 }
