@@ -7,9 +7,9 @@ import it.auties.protobuf.model.ProtobufType;
 import java.util.ArrayList;
 
 public record ModernBetaRepeatedMessage(
-        @ProtobufProperty(index = 1, type = ProtobufType.INT32, repeated = true)
+        @ProtobufProperty(index = 1, type = ProtobufType.INT32)
         ArrayList<Integer> content,
-        @ProtobufProperty(index = 2, type = ProtobufType.OBJECT, repeated = true)
+        @ProtobufProperty(index = 2, type = ProtobufType.OBJECT)
         ArrayList<ModernRepeatedMessage> content2
 ) implements ProtobufMessage {
 
